Created new test file for attempt failures
